requirements.txt
fastapi
uvicorn
python-dotenv
httpx
faiss-cpu
pypdf
tiktoken
